fix(cli): explicit-root JS files under node_modules keep their body read
`read_source_files`'s BFS applied the `maxNodeModuleJsDepth` skip gate
(`should_skip_js_in_node_modules`) uniformly to every pending path,
including explicit program roots (`files`/CLI args). A root JS file's
body was therefore never read, leaving it registered with zero
statements and starving every downstream CommonJS `module.exports`
export-surface computation for it — independent of `module`/
`moduleResolution` setting.

Oracle-verified against typescript@7.0.2: tsc never applies
`maxNodeModuleJsDepth` to an explicit root (it is a real program input,
not something reached by descending into node_modules), and correctly
infers the CJS export shape for such a root regardless of resolution
mode. This was previously masked by `any` under `commonjs`/classic
resolution and surfaced as a false TS2339 under Node16/NodeNext once
#16926 stopped emitting TS7016 for the same explicit-root case.
